Organize graph root modules
