+++ packages/extension/src/webview/components/graph/environment/browser.ts
