+++ packages/extension/src/webview/components/graph/environment/platform.ts
