+++ packages/extension/src/webview/components/graph/messages/listener.ts
@@ -1,6 +1,6 @@
-import type { ExtensionToWebviewMessage } from '../../../shared/protocol/extensionToWebview';
-import { getGraphWebviewMessageEffects, type GraphWebviewMessageEffect } from './messages/effects/routing';
-import type { FGNode } from './model/build';
+import type { ExtensionToWebviewMessage } from '../../../../shared/protocol/extensionToWebview';
+import { getGraphWebviewMessageEffects, type GraphWebviewMessageEffect } from './effects/routing';
+import type { FGNode } from '../model/build';
 
 export interface GraphMessageListenerOptions {
   applyEffects: (effects: GraphWebviewMessageEffect[]) => void;
