+++ packages/extension/src/webview/components/graph/view/sharedPropsOptions.ts
@@ -1,7 +1,7 @@
-import type { BuildSharedGraphPropsOptions } from './rendering/surface/sharedProps';
-import type { UseGraphInteractionRuntimeResult } from './runtime/use/interaction';
-import type { UseGraphStateResult } from './runtime/use/state';
-import type { DagMode } from '../../../shared/settings/modes';
+import type { BuildSharedGraphPropsOptions } from '../rendering/surface/sharedProps';
+import type { UseGraphInteractionRuntimeResult } from '../runtime/use/interaction';
+import type { UseGraphStateResult } from '../runtime/use/state';
+import type { DagMode } from '../../../../shared/settings/modes';
 
 export function buildGraphSharedPropsOptions({
   containerSize,
