+++ packages/extension/src/webview/components/graph/viewport/autoFit.ts
