+++ packages/extension/tests/webview/app/shell/mutations.test.tsx
@@ -20,7 +20,7 @@ const harness = vi.hoisted(() => ({
 
 const messageListeners: Array<(event: MessageEvent) => void> = [];
 
-vi.mock('../../../../src/webview/components/graph/view', () => ({
+vi.mock('../../../../src/webview/components/graph/view/component', () => ({
   default: (props: Record<string, unknown>) => {
     harness.graphProps = props;
     const data = props.data as { nodes: Array<{ id: string }>; edges: Array<{ id: string }> };
